added final 6 model variations
